requirement.txt
Django==1.5
MySQL-python==1.2.3
PIL==1.1.7
South==0.8.1
argparse==1.2.1
beautifulsoup4==4.3.2
django-debug-toolbar==1.2.1
django-registration==1.0
django-secretballot==0.2.4
git+https://github.com/sunlightlabs/django-secretballot.git
django-social-auth==0.7.7
httplib2==0.9
oauth2==1.5.211
python-openid==2.2.5
selenium==2.42.1
sqlparse==0.1.11
wsgiref==0.1.2
django-flaggit==0.0.3
django-simple-math-captcha==1.0.3
pytz
git+https://github.com/pcraston/django-tracking
hg+https://bitbucket.org/romanalexander/django-tracking
